backend/requirements.txt
fastapi
uvicorn
pydexcom
tconnectsync
sqlalchemy
mysql-connector-python
python-dotenv
apscheduler
setuptools<70
